수정 파일 읽고 insert 수정(FileInfoService)

+++ src/main/java/org/rb/sbsec/model/FileInfo.java
@@ -34,7 +34,7 @@ public class FileInfo implements Serializable {
 	private String fileName;
 	
 	private String fileExt;
-	private Integer fileSize;
+	private Long fileSize;
 	private String directory;
 	private String fileType;
 	private String delYn;
